Removed FilterList UI for now (collapsed into Settings component)
