--- src/devtools/views/Settings/FilterList.css
@@ -1,4 +1,0 @@
-.Filter {
-  display: block;
-  padding: 0 0 0.5rem;
-}
