+++ src/devtools/views/Settings/Settings.js
@@ -1,14 +1,16 @@
 // @flow
 
 import React, { useCallback, useContext, useMemo } from 'react';
+import { ElementTypeHostComponent } from 'src/types';
 import { useSubscription } from '../hooks';
 import { StoreContext } from '../context';
 import { SettingsContext } from './SettingsContext';
 import Store from 'src/devtools/store';
-import FilterList from './FilterList';
 import portaledContent from '../portaledContent';
 
 import styles from './Settings.css';
+
+import type { FilterPreferences } from 'src/types';
 
 function Settings(_: {||}) {
   const store = useContext(StoreContext);
@@ -42,6 +44,35 @@ function Settings(_: {||}) {
   );
   const collapseNodesByDefault = useSubscription<boolean, Store>(
     collapseNodesByDefaultSubscription
+  );
+
+  const filterPreferencesSubscription = useMemo(
+    () => ({
+      getCurrentValue: () => store.filterPreferences,
+      subscribe: (callback: Function) => {
+        store.addListener('filterPreferences', callback);
+        return () => store.removeListener('filterPreferences', callback);
+      },
+    }),
+    [store]
+  );
+  const filterPreferences = useSubscription<FilterPreferences, Store>(
+    filterPreferencesSubscription
+  );
+
+  const updateFilterPreferences = useCallback(
+    ({ currentTarget }) => {
+      const filterPreferences = store.filterPreferences;
+      if (currentTarget.checked) {
+        filterPreferences.hideElementsWithTypes.add(ElementTypeHostComponent);
+      } else {
+        filterPreferences.hideElementsWithTypes.delete(
+          ElementTypeHostComponent
+        );
+      }
+      store.filterPreferences = { ...filterPreferences };
+    },
+    [store]
   );
 
   const updateDisplayDensity = useCallback(
@@ -145,7 +176,16 @@ function Settings(_: {||}) {
           Collapse newly added components by default
         </label>
 
-        <FilterList />
+        <label className={styles.CheckboxOption}>
+          <input
+            type="checkbox"
+            checked={filterPreferences.hideElementsWithTypes.has(
+              ElementTypeHostComponent
+            )}
+            onChange={updateFilterPreferences}
+          />{' '}
+          Hide host components (e.g. <code>&lt;div&gt;</code>)
+        </label>
       </div>
 
       {store.supportsCaptureScreenshots && (
